@@ -21,6 +21,8 @@ using Address = System.UInt64;
 using Microsoft.Diagnostics.Utilities;
 using Microsoft.Diagnostics.HeapDump;
 using Azure.Core;
+using Azure.Identity;
+
 
 
 #if CROSS_GENERATION_LIVENESS
@@ -38,10 +40,10 @@ public class GCHeapDumper
     /// to dump a heap.  
     /// </summary>
     /// <param name="log"></param>
-    public GCHeapDumper(TextWriter log, TokenCredential symbolServerAuthCredential = null)
+    public GCHeapDumper(TextWriter log)
     {
         m_origLog = log;
-        m_symbolServerAuthCredential = symbolServerAuthCredential;
+        SymbolsAuthTokenCredential = new InteractiveBrowserCredential();
         m_copyOfLog = new StringWriter();
         m_log = new TeeTextWriter(m_copyOfLog, m_origLog);
 
@@ -284,11 +286,11 @@ public class GCHeapDumper
         {
             try
             {
-                dataTarget = DataTarget.CreateSnapshotAndAttach(processID, m_symbolServerAuthCredential);
+                dataTarget = DataTarget.CreateSnapshotAndAttach(processID, SymbolsAuthTokenCredential);
             }
             catch
             {
-                dataTarget = DataTarget.AttachToProcess(processID, Freeze, m_symbolServerAuthCredential);
+                dataTarget = DataTarget.AttachToProcess(processID, Freeze, SymbolsAuthTokenCredential);
             }
         }
         else
@@ -298,7 +300,7 @@ public class GCHeapDumper
                 UseOSMemoryFeatures = false // disable AWE
             };
 
-            dataTarget = DataTarget.LoadDump(processDumpFile, cacheOptions, m_symbolServerAuthCredential);
+            dataTarget = DataTarget.LoadDump(processDumpFile, cacheOptions, SymbolsAuthTokenCredential);
         }
 
         if (dataTarget.DataReader.PointerSize != IntPtr.Size)
@@ -390,6 +392,12 @@ public class GCHeapDumper
     /// The threshold at which we want to dump the heap when collecting cross-generation liveness data.
     /// </summary>
     public ulong PromotedBytesThreshold;
+
+
+    /// <summary>
+    /// The token credential to use for symbol server authentication.
+    /// </summary>
+    public TokenCredential SymbolsAuthTokenCredential;
 
     /// <summary>
     /// Force a .NET GC on a particular process. 
@@ -1578,8 +1586,6 @@ public class GCHeapDumper
     private TextWriter m_log;               // Where we send messages
     private StringWriter m_copyOfLog;       // We keep a copy of all logged messages here to append to output file. 
     private Stopwatch m_sw;                 // We keep track of how long it takes.  
-
-    private TokenCredential m_symbolServerAuthCredential;
 
     private GCHeapDump m_gcHeapDump;        // The image of what we are putting in the file
     private NodeIndex m_JSRoot = NodeIndex.Invalid;     // The root of the JS heap
