@@ -636,7 +636,10 @@ internal static class EventSourceFinder
         }
 
         // Compute the Sha1 hash
-        var sha1 = System.Security.Cryptography.SHA1.Create(); // lgtm [cs/weak-crypto]
+        // CodeQL [SM02196] The EventSource name to GUID protocol requires a SHA1 hash.
+        // CodeQL [SM03938] The EventSource name to GUID protocol requires a SHA1 hash.
+        // CodeQL [SM03939] The EventSource name to GUID protocol requires a SHA1 hash.
+        var sha1 = System.Security.Cryptography.SHA1.Create();
         byte[] hash = sha1.ComputeHash(bytes);
 
         // Create a GUID out of the first 16 bytes of the hash (SHA-1 create a 20 byte hash)
